refactor(er-matcher): extract fs_scorer module + cover separation gate
Move the goldenmatch-coupled FS scorer/blocker/cache layer out of build_corpus.py
into a new fs_scorer.py (build_corpus was 68% FS code while its docstring still
called it a stdlib-only blender). build_corpus keeps _fs_enrich_source as the
orchestrator and imports what it needs. All goldenmatch/polars imports stay
function-local, so fs_scorer is box-safe at import.

Add box tests for _check_separation (the posterior-vs-weighted gate -- a
margin/in-range regression silently flips every source to the weighted fallback):
in-range + margin-cleared passes; out-of-[0,1] fails; margin-not-met fails;
empty-class fails. Plus a round-trip for the pure cache helpers
(_pairs_hash/_save_fs_cache/_load_fs_cache). No goldenmatch (injected fake scorer).

Polish: module-level _DEFAULT_TAU replaces the bare 0.5 literals; posterior
signature computed once in _make_fs_scorer; _block_row_ids logs its BlockResult
shape once (T7 can confirm which goldenmatch bound); the best-effort excepts in
_threshold_from_config and _scorer_cfg_signature now log instead of swallowing.
