+++ scripts/er_matcher/test_build_corpus_fs.py
@@ -1,28 +1,33 @@
-"""Box tests for the PURE adapters in `build_corpus`'s FS-enrichment path
-(honest-yardstick Task 6): `_pairs_from_blocks` (block-pair -> gold-tagged
+"""Box tests for the PURE adapters in the FS-enrichment layer (`fs_scorer`,
+honest-yardstick Task 6): `_pairs_from_blocks` (block-pair -> gold-tagged
 candidate), `_gold_components` (match-row union-find + transitive closure),
-`_rebalance_negatives` (swap easy synth negatives for hard mined ones), and
-`_sample_labeled_pairs`.
+`_rebalance_negatives` (swap easy synth negatives for hard mined ones),
+`_sample_labeled_pairs`, `_check_separation` (the posterior-vs-weighted gate),
+and the disk-cache helpers (`_pairs_hash` / `_save_fs_cache` / `_load_fs_cache`).
 
-Box-safe: these functions are stdlib-only, so importing `build_corpus` here
-touches NEITHER goldenmatch NOR polars NOR yaml (every heavy import in
-`build_corpus` is function-local). No `--fs-enrich` codepath runs -- the
-goldenmatch-backed halves (`_fit_fs_posterior`, `_make_candidates`, ...) are
-verified by a real corpus build, not here."""
+Box-safe: `fs_scorer` keeps every goldenmatch/polars import function-local, so
+importing it here touches NEITHER goldenmatch NOR polars NOR yaml. No
+`--fs-enrich` codepath runs -- the goldenmatch-backed halves (`_fit_fs_posterior`,
+`_make_candidates`, ...) are verified by a real corpus build, not here."""
 
 from __future__ import annotations
 
 import os
 import sys
+from pathlib import Path
 
 sys.path.insert(0, os.path.dirname(__file__))
 
-from build_corpus import (  # noqa: E402
+from fs_scorer import (  # noqa: E402
+    _check_separation,
     _gold_components,
+    _load_fs_cache,
     _pairs_from_blocks,
+    _pairs_hash,
     _rebalance_negatives,
     _record_signature,
     _sample_labeled_pairs,
+    _save_fs_cache,
 )
 
 
@@ -40,7 +45,6 @@ def test_pairs_from_blocks_enumerates_within_block_pairs_and_tags_gold():
         "2": _rec("bob", "e2"),
     }
     # One block of all three ids; gold links 0<->1 only.
-    gold = {("0", "1")}
     gold_linked = lambda a, b: (a["eid"], b["eid"]) == ("e1", "e1")  # noqa: E731
 
     out = _pairs_from_blocks([["0", "1", "2"]], records_by_id, gold_linked)
@@ -49,7 +53,6 @@ def test_pairs_from_blocks_enumerates_within_block_pairs_and_tags_gold():
     assert pairs == {("0", "1"): True, ("0", "2"): False, ("1", "2"): False}
     # every candidate carries the actual record dicts for scoring
     assert all("a" in c and "b" in c for c in out)
-    del gold
 
 
 def test_pairs_from_blocks_dedups_across_passes_canonical_order():
@@ -148,3 +151,72 @@ def test_sample_labeled_pairs_balances_and_tags_match_flag():
     assert sum(is_match) == 5
     assert sum(1 for m in is_match if not m) == 3
     assert all(isinstance(a, dict) and isinstance(b, dict) for a, b, _m in sample)
+
+
+# ── _check_separation (the posterior-vs-weighted gate) ──────────────────────
+
+
+def _sample(scores):
+    """A labeled sample as (a, b, is_match) where scorer reads a['s'] as the
+    injected score -- lets these tests drive `_check_separation` with no
+    goldenmatch."""
+    return [({"s": s}, {"s": s}, is_m) for s, is_m in scores]
+
+
+def _fake_scorer(a, b):
+    return a["s"]
+
+
+def test_check_separation_passes_when_margin_cleared_and_in_range():
+    sample = _sample([(0.95, True), (0.9, True), (0.05, False), (0.1, False)])
+    ok, mean_m, mean_n, in_range = _check_separation(_fake_scorer, sample)
+    assert ok is True
+    assert in_range is True
+    assert mean_m > mean_n
+
+
+def test_check_separation_fails_when_out_of_unit_range():
+    # match mean beats non-match mean by a mile, but a score sits outside [0,1]
+    sample = _sample([(1.4, True), (0.9, True), (0.05, False)])
+    ok, _mean_m, _mean_n, in_range = _check_separation(_fake_scorer, sample)
+    assert in_range is False
+    assert ok is False
+
+
+def test_check_separation_fails_when_margin_not_met():
+    # both classes ~0.5: mean(match) does NOT beat mean(non-match) by the margin
+    sample = _sample([(0.52, True), (0.5, True), (0.5, False), (0.48, False)])
+    ok, mean_m, mean_n, in_range = _check_separation(_fake_scorer, sample)
+    assert in_range is True
+    assert mean_m - mean_n < 0.1  # below _SEPARATION_MARGIN
+    assert ok is False
+
+
+def test_check_separation_fails_when_a_class_is_empty():
+    # no non-match scores at all -> cannot verify separation -> not ok
+    sample = _sample([(0.9, True), (0.95, True)])
+    ok, _mean_m, _mean_n, _in_range = _check_separation(_fake_scorer, sample)
+    assert ok is False
+
+
+# ── pure disk-cache helpers (round-trip) ────────────────────────────────────
+
+
+def test_pairs_hash_deterministic_and_order_sensitive():
+    p1 = [{"a": {"n": "x"}, "b": {"n": "y"}, "label": "match", "eid_a": "0", "eid_b": "1"}]
+    p2 = [{"a": {"n": "x"}, "b": {"n": "z"}, "label": "match", "eid_a": "0", "eid_b": "1"}]
+    assert _pairs_hash(p1) == _pairs_hash(p1)  # stable
+    assert _pairs_hash(p1) != _pairs_hash(p2)  # content-sensitive
+
+
+def test_fs_cache_save_load_round_trip(tmp_path: Path):
+    key = "deadbeef"
+    payload = {
+        "train": [{"label": "match", "confidence": 0.97}],
+        "val": [],
+        "test": [{"label": "no_match", "confidence": 0.03, "negative_kind": "fs_mined"}],
+    }
+    assert _load_fs_cache(tmp_path, key) is None  # miss before write
+    _save_fs_cache(tmp_path, key, payload)
+    assert _load_fs_cache(tmp_path, key) == payload  # hit after write
+    assert (tmp_path / ".fs_cache" / f"{key}.json").exists()
